+++ packages/functions/prod/src/function/index.ts
@@ -3,3 +3,34 @@ export { HandlerEvent } from './handler_event.js'
 export { BuilderHandler, Handler, BackgroundHandler, HandlerCallback, StreamingHandler } from './handler.js'
 export { BuilderResponse, HandlerResponse, StreamingResponse } from './handler_response.js'
 export { Context, Config } from './v2.js'
+export { NetlifyFunction } from './netlify_function.js'
+
+export type {
+  Deploy,
+  DeploySite,
+  DeployBuildingEvent,
+  DeploySucceededEvent,
+  DeployFailedEvent,
+  DeployDeletedEvent,
+  DeployLockedEvent,
+  DeployUnlockedEvent,
+  DeployBuildingHandler,
+  DeploySucceededHandler,
+  DeployFailedHandler,
+  DeployDeletedHandler,
+  DeployLockedHandler,
+  DeployUnlockedHandler,
+  User,
+  UserLoginEvent,
+  UserSignupEvent,
+  UserValidateEvent,
+  UserModifiedEvent,
+  UserDeletedEvent,
+  UserLoginHandler,
+  UserSignupHandler,
+  UserValidateHandler,
+  UserModifiedHandler,
+  UserDeletedHandler,
+  FormSubmittedEvent,
+  FormSubmittedHandler,
+} from '@netlify/types'