+++ packages/types/src/main.ts
@@ -1,8 +1,36 @@
 export type { Context } from './lib/context/context.js'
 export type { Cookie } from './lib/context/cookies.js'
-export type { DeployEvent } from './lib/events/deploy.js'
-export type { IdentityEvent } from './lib/events/identity.js'
-export type { FormSubmissionEvent } from './lib/events/submission.js'
+export type {
+  DeployEvent,
+  Deploy,
+  DeploySite,
+  DeployBuildingEvent,
+  DeploySucceededEvent,
+  DeployFailedEvent,
+  DeployDeletedEvent,
+  DeployLockedEvent,
+  DeployUnlockedEvent,
+  DeployBuildingHandler,
+  DeploySucceededHandler,
+  DeployFailedHandler,
+  DeployDeletedHandler,
+  DeployLockedHandler,
+  DeployUnlockedHandler,
+} from './lib/events/deploy.js'
+export type {
+  User,
+  UserLoginEvent,
+  UserSignupEvent,
+  UserValidateEvent,
+  UserModifiedEvent,
+  UserDeletedEvent,
+  UserLoginHandler,
+  UserSignupHandler,
+  UserValidateHandler,
+  UserModifiedHandler,
+  UserDeletedHandler,
+} from './lib/events/identity.js'
+export type { FormSubmittedEvent, FormSubmittedHandler } from './lib/events/submission.js'
 export type { EnvironmentVariables } from './lib/environment-variables.js'
 export type { NetlifyGlobal } from './lib/globals.js'
 export type { Site } from './lib/context/site.js'