+++ packages/cache/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [3.4.4](https://github.com/netlify/primitives/compare/cache-v3.4.3...cache-v3.4.4) (2026-03-19)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @netlify/dev-utils bumped from 4.4.2 to 4.4.3
+    * @netlify/types bumped from 2.5.0 to 2.6.0
+
+## [3.4.3](https://github.com/netlify/primitives/compare/cache-v3.4.2...cache-v3.4.3) (2026-03-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @netlify/dev-utils bumped from 4.4.1 to 4.4.2
+    * @netlify/types bumped from 2.4.0 to 2.5.0
+
 ## [3.4.2](https://github.com/netlify/primitives/compare/cache-v3.4.1...cache-v3.4.2) (2026-03-12)